strongly type LSP request
